move map data location
