+++ .prettierignore
@@ -1,4 +1,3 @@
 # Ignore json files in the data folder
-/src/data/
-/src/data/100000_full.json
+/src/_data/
 /public/assets/lib/detection.js
