+++ src/pages/blog/_posts/map-maintenance-guide.mdx
@@ -6,9 +6,9 @@ title: 就诊地图维护指南
 
 就诊地图数据位于：
 
-- `src/data/medicalData.json` : 国内版
-- `src/data/medicalChildData.json`: 儿童版
-- `src/data/medicalAbroadData.json`: 国际版
+- `src/_data/medicalData.json` : 国内版
+- `src/_data/medicalChildData.json`: 儿童版
+- `src/_data/medicalAbroadData.json`: 国际版
 
 ## 地图数据结构
