+++ src/pages/quotients/eq60.jsx
@@ -5,7 +5,7 @@ import QuestionItem from "@/components/QuestionItem";
 import QuestionResult from "@/components/QuestionResult";
 import QuestionInfo from "@/components/QuestionInfo";
 import QuestionInfoAlert from "@/components/QuestionInfoAlert";
-import questionData from "@/data/questionEQ60.json";
+import questionData from "@/_data/questionEQ60.json";
 import BackToTop from "@/components/BackToTop";
 
 class EQ60 extends Component {
